+++ R/getThemes.R
@@ -27,5 +27,10 @@ getThemes <- function(key = getOption('brickset_key'),
 	}
 
 	themes_json <- jsonlite::fromJSON(httr::content(themes, as = 'text', encoding = "UTF-8"))
+
+	if(themes_json$status == 'error') {
+		stop(paste0('Error getting sets: ', themes_json$message))
+	}
+
 	return(themes_json[[3]])
 }