+++ NEWS.md
@@ -1,6 +1,7 @@
 # brickset (development version)
 
 * Added number, thumbnailURL, and imageURL to the data.frame returned by the getSets function.
+* Added check for errors in returned JSON. For example, exceeding the API limit will return a user friendly message.
 * Added code of conduct.
 * Fixed location of README figures.