+++ NEWS.md
@@ -1,5 +1,6 @@
-# brickset (development version)
+# brickset 2024.0.0
 
+* Added 2023 sets.
 * Added number, thumbnailURL, and imageURL to the data.frame returned by the getSets function.
 * Will check for errors in returned JSON. For example, exceeding the API limit will return a user friendly message.
 * Added code of conduct.